BC25 Preview Scaffolding
Added base files for BC25 preview, based on WallpaperPreviewActivity
preview flow. These files are simplified versions of the
WallpaperPreviewActivity files, moved into the "common" package so they
can be shared between com.android.wallpaper.picker's customization
package and preview package. To keep the CL leaner, the ability to
preview the workspace, preview colors and to set crophints are not yet
migrated over and will be added in future CLs. Also the ability to
preview different home and lock wallpapers when previewing current
wallpapers is yet to be added. Currently to test the flow, the home
wallpaper is set as the wallpaper model.

Flag: com.android.wallpaper.new_picker_ui_flag
Bug: 348462236
Test: manually verified
Test: unit tests added
Change-Id: I29f4ef2e09b341594ef9ac3f630a7027a145e51b