Implement view model navigation for Categories
This CL implements the view model logic for navigation for the category screen.
The goal is to encapsulate the business logic for navigation in the view model. This will lead to better
separation of concern and  will allow the navigation logic to be tested, as i have added a few nav tests to

Flag: com.android.wallpaper.refactor_wallpaper_category_flag
Bug: 350090051
Test: manual
Change-Id: Ibba9c754c173acf837c5730d9d6771b76d5e1bcc