Enable wallpaper surface to resize to surface view
Enable a fixed surface size and use Hardware Scaler to scale the surface
instead. This resolves the issue with both live and static wallpaper
surfaces never resizing according to the surface view size. The caveat
with this solution is that if the surface view scales up, the resolution
of the surface will be sub-par. However, in our use case, the surface
is only scaled down, never scaled up, so the resolution should not be an
issue.

Flag: com.android.wallpaper.new_picker_ui_flag
Bug: 348462236
Test: maually verified with live and static wallpapers
Change-Id: I5450baabc89d10d65b14288164d189d6bae34b54