Enable navigation from new picker activity to multi-crop activity
Enable navigation to multi-crop preview activity by tapping on new
picker preview. Enable passing of WallpaperModel between activities by
storing it in an application-scoped repository. This is a workaround to
the fact that WallpaperModel cannot be easily made a Parcelable, because
it uses the Asset class.

Flag: com.android.wallpaper.new_picker_ui_flag
Bug: 348462236
Test: manually verified by tapping on preview
Change-Id: Ib229d6f83ca0b4f646bb56ba67d52631d7b79b68