Update More Wallpapers text color on system color change (1/2)
In addition, add a shouldAnimate function parameter to the
ColorUpdateBinder, to allow specifying whether a color update should be
animated. This is generally based on whether the view is showing, which
can be determined based on which screen or customization option is
selected in the view model.

Flag: com.android.systemui.new_picker_ui
Bug: 299510645
Test: manually verified More Wallpapers color updates with no restart
Change-Id: Id38a9346339a83b94e5bfdb81a1b5cdede419a55